refactor(api): extract scan routes and utilities
